+++ util/math/vector3i.h
@@ -360,6 +360,10 @@ inline Vector3i operator>>(const Vector3i &a, int b) {
 	return Vector3i(a.x >> b, a.y >> b, a.z >> b);
 }
 
+inline Vector3i operator&(const Vector3i &a, uint32_t b) {
+	return Vector3i(a.x & b, a.y & b, a.z & b);
+}
+
 #endif // VOXEL_CUSTOM_VECTOR3I
 
 // For Godot